made the edits so the thing works with ubuntu

+++ reddit.py
@@ -20,10 +20,10 @@ def make_square(img_path, min_size=256, fill_color=(0, 0, 0, 0)):
 	im = Image.open(img_path)
 	x, y = im.size
 	size = max(min_size, x, y)
-	new_im = Image.new('RGBA', (size, size), fill_color)
+	new_im = Image.new('RGB', (size, size), fill_color)
 	new_im.paste(im, (int((size - x) / 2), int((size - y) / 2)))
+	img_path = img_path[:-3] + "jpg"
 	new_im.save(img_path)
-
 
 rc = praw.Reddit(client_id="362R_F8jZdyhiw",
 			 client_secret=clientscrt,
@@ -32,32 +32,32 @@ rc = praw.Reddit(client_id="362R_F8jZdyhiw",
 
 subreddit = rc.subreddit("memes")
 caption_bottom = "\n\n\n\n\n\n #meme #memes #funny #dankmemes #memesdaily #funnymemes #lol #dank #humor #follow #like #dankmeme #lmao #love #anime #ol #edgymemes #dailymemes #comedy #instagram #fun #offensivememes #fortnite #funnymeme #tiktok #memer #memez #memepage #memestagram #bhfyp"
-base = "F:/bot/Unexpected/"
-uploaded_number=9
+base = "/home/akshat"
+uploaded_number=0
 
 instagram_client = Instagram(user, passwrd)
 
 
 while 1:
-	if datetime.now().hour >= 9:
+	if 1:
 		cont = 0
 		for sub in subreddit.top('day'):
 			if cont >= 1: break
 
 			id=sub.url.split("/")[-1]
 			url = sub.url
+			path = os.path.join(base,f"{id}")
+			id = id[:-3] + "jpg"
 
 			if not sub.stickied and not_uploaded(id):
 
-				path = os.path.join(base,f"{id}")
 				req = requests.get(url)
-
 				if req.ok:
 					with open(path, "wb") as f:
 						f.write(req.content)
 
 					make_square(path)
-					instagram_client.upload(sub.title+caption_bottom, path)
+					instagram_client.upload(sub.title+caption_bottom, id)
 
 					with open("record.txt", "a+") as f:
 						f.write(id+"\n")
@@ -67,5 +67,6 @@ while 1:
 
 					if uploaded_number >=10:
 						for file in os.listdir(base):
-							os.remove(os.path.join(base,file))
+							if file.endswith("jpg"):
+								os.remove(os.path.join(base,file))
 	sleep(0.5 * 60*60)
